A ton of changes
7/25/18 9:05PM
- I experemented a lot, broke some things, added some things, removed some things, hair turned gray....
- Too confusing to try to list what I did...

+++ windows/win32Platform.cpp
@@ -142,7 +142,7 @@ RESULT Win32ReadFileToBuffer(char* buffer, const uint64 bufferSize, const uint64
 
 	DWORD bytesRead = 0;
 
-	if (!ReadFile(file, buffer + offset, bytes, &bytesRead, NULL))
+	if (!ReadFile(file, buffer + offset, (DWORD)bytes, &bytesRead, NULL))
 	{
 		DWORD error = GetLastError();
 		LPTSTR errorText = NULLPTR;
